ractor: SessionsRepositoryをDIしてAPIリクエストするように変更
