+++ lib/models/controllers/controllers.dart
@@ -1,0 +1,1 @@
+export 'sessions_controller/sessions_controller.dart';
