+++ lib/pages/home_pages/twitter_login_button.dart
@@ -5,6 +5,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:twitter_login/twitter_login.dart';
+import 'package:event_follow/models/models.dart';
 
 import '../../main.dart';
